Consolidate translators
