+++ src/Internal/Query/ParadeDbMethodCallTranslatorPlugin.cs
@@ -1,6 +1,5 @@
 using Microsoft.EntityFrameworkCore.Query;
 using Microsoft.EntityFrameworkCore.Storage;
-using ParadeDB.EntityFrameworkCore.Internal.Query.Translators;
 
 namespace ParadeDB.EntityFrameworkCore.Internal.Query;
 
@@ -11,18 +10,7 @@ internal sealed class ParadeDbMethodCallTranslatorPlugin : IMethodCallTranslator
         IRelationalTypeMappingSource typeMappingSource
     )
     {
-        Translators =
-        [
-            new ModifierTranslator(sqlExpressionFactory),
-            new OperatorTranslator(sqlExpressionFactory),
-            new ScoreTranslator(sqlExpressionFactory),
-            new SnippetTranslator(sqlExpressionFactory),
-            new ProximityTranslator(sqlExpressionFactory),
-            new TokenizeTranslator(sqlExpressionFactory),
-            new MoreLikeThisTranslator(sqlExpressionFactory),
-            new AliasTranslator(sqlExpressionFactory),
-            new AggregateTranslator(sqlExpressionFactory, typeMappingSource),
-        ];
+        Translators = [new Translator(sqlExpressionFactory, typeMappingSource)];
     }
 
     public IEnumerable<IMethodCallTranslator> Translators { get; }
