+++ tests/QueryTests.cs
@@ -586,7 +586,7 @@ public sealed class MatchAllTests : TestBase
         await using var context = DbFixture.CreateContext();
 
         var query = context.MockItems.Where(p =>
-            EF.Functions.Match(p.Description, Pdb.Proximity("sleek").Within(1, "shoes"))
+            EF.Functions.Proximity(p.Description, Pdb.Proximity("sleek").Within(1, "shoes"))
         );
 
         var sql = """
@@ -605,7 +605,7 @@ public sealed class MatchAllTests : TestBase
         await using var context = DbFixture.CreateContext();
 
         var query = context.MockItems.Where(p =>
-            EF.Functions.Match(
+            EF.Functions.Proximity(
                 p.Description,
                 Pdb.Proximity("sleek").Within(1, "shoes", ordered: true)
             )
@@ -631,7 +631,7 @@ public sealed class MatchAllTests : TestBase
         int distance = 1;
 
         var query = context.MockItems.Where(p =>
-            EF.Functions.Match(p.Description, Pdb.Proximity(left).Within(distance, right))
+            EF.Functions.Proximity(p.Description, Pdb.Proximity(left).Within(distance, right))
         );
 
         var sql = """
@@ -653,7 +653,7 @@ public sealed class MatchAllTests : TestBase
         await using var context = DbFixture.CreateContext();
 
         var query = context.MockItems.Where(p =>
-            EF.Functions.Match(p.Description, Pdb.ProximityRegex("sl.*").Within(1, "shoes"))
+            EF.Functions.Proximity(p.Description, Pdb.ProximityRegex("sl.*").Within(1, "shoes"))
         );
 
         var sql = """
@@ -674,7 +674,7 @@ public sealed class MatchAllTests : TestBase
         string pattern = "sl.*";
 
         var query = context.MockItems.Where(p =>
-            EF.Functions.Match(p.Description, Pdb.ProximityRegex(pattern).Within(1, "shoes"))
+            EF.Functions.Proximity(p.Description, Pdb.ProximityRegex(pattern).Within(1, "shoes"))
         );
 
         var sql = """
@@ -694,7 +694,10 @@ public sealed class MatchAllTests : TestBase
         await using var context = DbFixture.CreateContext();
 
         var query = context.MockItems.Where(p =>
-            EF.Functions.Match(p.Description, Pdb.ProximityRegex("sl.*", 100).Within(1, "shoes"))
+            EF.Functions.Proximity(
+                p.Description,
+                Pdb.ProximityRegex("sl.*", 100).Within(1, "shoes")
+            )
         );
 
         var sql = """
@@ -713,7 +716,7 @@ public sealed class MatchAllTests : TestBase
         await using var context = DbFixture.CreateContext();
 
         var query = context.MockItems.Where(p =>
-            EF.Functions.Match(
+            EF.Functions.Proximity(
                 p.Description,
                 Pdb.ProximityArray("sleek", "white").Within(1, "shoes")
             )
@@ -738,7 +741,7 @@ public sealed class MatchAllTests : TestBase
         string t2 = "white";
 
         var query = context.MockItems.Where(p =>
-            EF.Functions.Match(p.Description, Pdb.ProximityArray(t1, t2).Within(1, "shoes"))
+            EF.Functions.Proximity(p.Description, Pdb.ProximityArray(t1, t2).Within(1, "shoes"))
         );
 
         var sql = """
@@ -759,7 +762,7 @@ public sealed class MatchAllTests : TestBase
         await using var context = DbFixture.CreateContext();
 
         var query = context.MockItems.Where(p =>
-            EF.Functions.Match(
+            EF.Functions.Proximity(
                 p.Description,
                 Pdb.ProximityArray(Pdb.ProximityRegex("sl.*"), Pdb.ProximityArray("white"))
                     .Within(1, "shoes")
@@ -782,7 +785,7 @@ public sealed class MatchAllTests : TestBase
         await using var context = DbFixture.CreateContext();
 
         var query = context.MockItems.Where(p =>
-            EF.Functions.Match(
+            EF.Functions.Proximity(
                 p.Description,
                 Pdb.Proximity("sleek").Within(1, "running").Within(2, "shoes")
             )
